Changed Variables to VEX Disc Specifications
Changed the mass & area to that of a vex disc (as listed in the game manual). Also changed initial launch velocity to our disc launcher v0.2's theoretical initial launch velocity. The desired coordinate was changed to 14 feet away, & to the height of the middle of a high goal. Result was a 5.617 m/s initial velocity.