Increased Precision & Tuned Kp
Increased the precision of the simulation & PID by increasing deltaT & decreasing maxError. Also tuned Kp by running the program to completion, then modifying the Kp's values, and checking for a positive or negative result in the number of internal iterations after running it again; this process was continued until any major change increased the number of internal iterations.