Fixed Alpha
Alpha is the angle of attack, and the angle of attack is supposed to be relative to the velocity. Previously it was relative to the ground, now it is updated every iteration to be relative to the angle of the velocity. The change produced by this appeared to be minimal, but it should be more correct.